Move column width functions into table helper
